sum += (long) Math.pow(i, 3);

+++ src/main/java/ru/job4j/codewars/SumCubes.java
@@ -5,7 +5,7 @@ public class SumCubes {
         System.out.println(n);
         long sum = 0;
         for (int i = 0; i <= n; i++) {
-            sum += Math.pow(i, 3);
+            sum += (long) Math.pow(i, 3);
         }
         return sum;
